Add suspected patterns and literals
As per feedback from Gabe, use similar logic to identify offending
cookies. This is a direct port of the PHP implemented in constants.
If these values change frequently, maybe modifying the Cookie Cutter
repo to have a config file that uses a parse able format may be worth
exploring.

Rename variables to longer more communicative names (e.g., `l` =>
`line`) to make the script easier to read now that it may be used
outside of a proof of concept.