Add colorized icon terminal output
Move from a "checking" to "checked" message per site in the terminal.
This allows for including a status and offending cookie count in the
output while still allowing for a message with "Y of X" sites.

Just in case someone despises icons in the terminal, an option for
`--no-color` was added.