Add thread safe pool for parallel checks
Add a `pool` option that defaults to 4. If a user wants to run serial
scans the pool size can be set to 1 and it will be effectively
synchronous.

The side effects here are that the sites may not finish scanning in the
same order each time resulting in different console output and CSV
ordering. The console output was modified to print the "Y of X"
statement in the order of completion rather than the sites order in the
`data/sites.txt` input. This allows for an ordered numbering while
maintaining the speed gained with using threads.

The Mutex is used to ensure that the CSV gets all the data in a thread
safe manner. This can lead to the spreadsheet/CSV changing order per
report run. This side effect is easily fixed by opening the CSV file and
ordering the resulting rows using the users favorite CSV viewer.

Adding dynamic pooling significantly reduces the scan times of a longer
list of URLs and paves the way for Gabe's hope to run the tool on
potentially several thousand sites.